Migrate to features folder structure
