+++ lib/core/db/nouns_database/enums/level.dart
