+++ lib/core/db/nouns_database/utils/equal_list.dart
