+++ lib/core/db/nouns_database/utils/open_database_stub.dart
