+++ lib/core/db/nouns_database/utils/open_web_database.dart
@@ -1,0 +1,7 @@
+import 'package:der_die_das/core/db/nouns_database/utils/import_from_assets.dart';
+import 'package:drift/drift.dart';
+import 'package:drift/web.dart';
+
+QueryExecutor openDatabase() {
+  return WebDatabase('db', initializer: importDbFromAssets);
+}
