+++ lib/core/db/nouns_database/utils/platform_open_database.dart
