+++ lib/core/extensions/normaize_string_extensions.dart
